+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "datasets-viewer" extension will be documented in thi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [Unreleased]
+
+### Miscellaneous
+
+- Update dependencies
+
 ## [0.0.3] - 2024-03-27
 
 ### Added